GH-2501: Fix Class Tangle
See https://github.com/spring-projects/spring-kafka/issues/2501

New tangle between `container <-> properties`.

Move the new properties and accessors to the abstract container.
