+++ spring-kafka-docs/src/main/asciidoc/kafka.adoc
@@ -1749,7 +1749,7 @@ For the `ConcurrentMessageListenerContainer`, the `<beanName>` part of the threa
 So, with a bean name of `container`, threads in this container will be named `container-0-C-1`, `container-1-C-1` etc., after the container is started the first time; `container-0-C-2`, `container-1-C-2` etc., after a stop and subsequent start.
 
 Starting with version `3.0.1`, you can now change the name of the thread, regardless of which executor is used.
-Set the `ContainerProperties.changeConsumerThreadName` property to `true` and the `ContainerProperties.threadNameSupplier` will be invoked to obtain the thread name.
+Set the `AbstractMessageListenerContainer.changeConsumerThreadName` property to `true` and the `AbstractMessageListenerContainer.threadNameSupplier` will be invoked to obtain the thread name.
 This is a `Function<MessageListenerContainer, String>`, with the default implementation returning `container.getListenerId()`.
 
 [[kafka-listener-meta]]
