+++ spring-kafka/src/main/java/org/springframework/kafka/listener/KafkaMessageListenerContainer.java
@@ -1397,9 +1397,10 @@ public class KafkaMessageListenerContainer<K, V> // NOSONAR line count
 		}
 
 		protected void initialize() {
-			if (this.containerProperties.isChangeConsumerThreadName()) {
+			if (KafkaMessageListenerContainer.this.thisOrParentContainer.isChangeConsumerThreadName()) {
 				Thread.currentThread().setName(
-						this.containerProperties.getThreadNameSupplier().apply(KafkaMessageListenerContainer.this));
+						KafkaMessageListenerContainer.this.thisOrParentContainer.getThreadNameSupplier()
+								.apply(KafkaMessageListenerContainer.this));
 			}
 			publishConsumerStartingEvent();
 			this.consumerThread = Thread.currentThread();
